cdk-snowflake/requirements.txt
# AWS CDK and core dependencies
aws-cdk-lib==2.211.0
constructs>=10.0.0

# Snowflake connector
snowflake-connector-python[pandas]==3.12.2

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
